refactor show tags
