+++ mods/tql/tql_test.go
@@ -302,7 +302,7 @@ func TestDatabaseTql(t *testing.T) {
 				CSV(header(true))
 				`,
 			ExpectCSV: []string{
-				"_ID,NAME,ROW_COUNT,MIN_TIME,MAX_TIME,RECENT_ROW_TIME,MIN_VALUE,MIN_VALUE_TIME,MAX_VALUE,MAX_VALUE_TIME",
+				"ID,NAME,ROW_COUNT,MIN_TIME,MAX_TIME,RECENT_ROW_TIME,MIN_VALUE,MIN_VALUE_TIME,MAX_VALUE,MAX_VALUE_TIME",
 				"1,tag1,2,1692686707380411000,1692686708380411000,1692686708380411000,NULL,NULL,NULL,NULL",
 				"\n",
 			},
