--- spi/do_license.go
@@ -1,59 +1,0 @@
-package spi
-
-import (
-	"context"
-	"errors"
-	"fmt"
-	"os"
-	"strings"
-	"time"
-
-	"github.com/machbase/neo-client/api"
-)
-
-type LicenseInfo struct {
-	Id            string `json:"id"`
-	Type          string `json:"type"`
-	Customer      string `json:"customer"`
-	Project       string `json:"project"`
-	CountryCode   string `json:"countryCode"`
-	InstallDate   string `json:"installDate"`
-	IssueDate     string `json:"issueDate"`
-	LicenseStatus string `json:"licenseStatus,omitempty"`
-}
-
-func GetLicenseInfo(ctx context.Context, conn api.Conn) (*LicenseInfo, error) {
-	ret := &LicenseInfo{}
-	var violateStatus int
-	row := conn.QueryRow(ctx, "select ID, TYPE, CUSTOMER, PROJECT, COUNTRY_CODE, INSTALL_DATE, ISSUE_DATE, VIOLATE_STATUS, VIOLATE_MSG from v$license_info")
-	if err := row.Scan(&ret.Id, &ret.Type, &ret.Customer, &ret.Project, &ret.CountryCode, &ret.InstallDate, &ret.IssueDate, &violateStatus, &ret.LicenseStatus); err != nil {
-		return nil, err
-	}
-	if violateStatus == 0 {
-		ret.LicenseStatus = "Valid"
-	}
-	return ret, nil
-}
-
-func InstallLicenseFile(ctx context.Context, conn api.Conn, path string) (*LicenseInfo, error) {
-	if strings.ContainsRune(path, ';') {
-		return nil, errors.New("invalid license file path")
-	}
-	result := conn.Exec(ctx, "alter system install license='"+path+"'")
-	if result.Err() != nil {
-		return nil, result.Err()
-	}
-	return GetLicenseInfo(ctx, conn)
-}
-
-func InstallLicenseData(ctx context.Context, conn api.Conn, licenseFilePath string, content []byte) (*LicenseInfo, error) {
-	_, err := os.Stat(licenseFilePath)
-	if err == nil {
-		// backup existing file
-		os.Rename(licenseFilePath, fmt.Sprintf("%s_%s", licenseFilePath, time.Now().Format("20060102_150405")))
-	}
-	if err := os.WriteFile(licenseFilePath, content, 0640); err != nil {
-		return nil, err
-	}
-	return InstallLicenseFile(ctx, conn, licenseFilePath)
-}
