+++ spi/show_test.go
@@ -450,11 +450,11 @@ func TestShowTables(t *testing.T) {
 			},
 		},
 		{
-			name:    "QueryTags",
-			fn:      func() spi.ResultSet { return spi.ResultSet(spi.QueryTags(t.Context(), conn, "rs_data", "test1")) },
-			columns: []string{"_ID", "NAME", "ROW_COUNT", "MIN_TIME", "MAX_TIME", "RECENT_ROW_TIME", "MIN_VALUE", "MIN_VALUE_TIME", "MAX_VALUE", "MAX_VALUE_TIME"},
+			name:    "ShowTags",
+			fn:      func() spi.ResultSet { return spi.ResultSet(spi.ShowTags(t.Context(), conn, "rs_data", "test1")) },
+			columns: []string{"ID", "NAME", "ROW_COUNT", "MIN_TIME", "MAX_TIME", "RECENT_ROW_TIME", "MIN_VALUE", "MIN_VALUE_TIME", "MAX_VALUE", "MAX_VALUE_TIME"},
 			expects: [][]any{
-				{int64(1), "test1", int64(2), parseTime("2024-01-01 00:00:00"), parseTime("2024-01-02 00:00:00"), parseTime("2024-01-02 00:00:00"), float64(1), parseTime("2024-01-01 00:00:00"), float64(2), parseTime("2024-01-02 00:00:00")},
+				{int64(1), "test1", int64(2), parseTime("2024-01-01 00:00:00.000"), parseTime("2024-01-02 00:00:00.000"), parseTime("2024-01-02 00:00:00"), float64(1), parseTime("2024-01-01 00:00:00.000"), float64(2), parseTime("2024-01-02 00:00:00.000")},
 			},
 		},
 	}
